+++ CHANGELOG.md
@@ -54,7 +54,10 @@
   * The public attribute `searchInputComponent` has been renamed to `quickFilterComponent`.
   * The output `onListDataLoaded` has been renamed to `onDataLoaded`.
   * The output `onPaginatedListDataLoaded` has been renamed to `onPaginatedDataLoaded`.
-* **o-table**: Attribute `show-table-buttons-text` has been renamed to `show-buttons-text`.
+* **o-table**:
+  * The attribute `show-table-buttons-text` has been renamed to `show-buttons-text`.
+  * The output `onTableDataLoaded` has been renamed to `onDataLoaded`.
+  * The output `onPaginatedTableDataLoaded` has been renamed to `onPaginatedDataLoaded`.
 * **o-form**: Output `onFormDataLoaded` has been renamed to `onDataLoaded`.
 * **o-user-info**: removed input `use-flag-icons`. This component doesn't contain the language selector anymore.
 * Changes in CSS classes naming: